Split unit tests up.
