+++ test/test_math/main.cpp
@@ -1,8 +1,9 @@
 #include <Arduino.h>
 #include <unity.h>
-
 #include "tests_crankmaths.h"
-#include "tests_maths.h"
+
+extern void testPercent(void);
+extern void testDivision(void);
 
 #define UNITY_EXCLUDE_DETAILS
 
@@ -17,7 +18,8 @@ void setup()
     UNITY_BEGIN();    // IMPORTANT LINE!
 
     testCrankMaths();
-    testMaths();
+    testPercent();
+    testDivision();
 
     UNITY_END(); // stop unit testing
 }
